random number generator moved to utils
